chore: :truck: move tests to match change in folder
