+++ tests/core/check_datapackage/test_check_error_matcher.py
@@ -1,7 +1,7 @@
 from pytest import mark
 
-from seedcase_sprout.core.checks.check_error import CheckError
-from seedcase_sprout.core.checks.check_error_matcher import CheckErrorMatcher
+from seedcase_sprout.core.check_datapackage.check_error import CheckError
+from seedcase_sprout.core.check_datapackage.check_error_matcher import CheckErrorMatcher
 
 
 @mark.parametrize(
