+++ tests/core/check_datapackage/test_get_full_json_path_from_error.py
@@ -1,7 +1,7 @@
 from jsonschema import ValidationError
 from pytest import mark
 
-from seedcase_sprout.core.checks.get_full_json_path_from_error import (
+from seedcase_sprout.core.check_datapackage.get_full_json_path_from_error import (
     get_full_json_path_from_error,
 )
